requirements.txt
streamlit
opencv-python
numpy
pillow
torch
torchvision

